* Define a non-zero FUDGE_FACTOR for GSSAPI; I have it set to 64 now, even though I tend to see 52 bytes of increased data because I'm not sure that the 52 bytes is constant across all implementations ands options.
* When allocating outbuf, set bufsize to the size that was actually
allocated; it tends not to be nbyte+FUDGE_FACTOR exactly, and you
smash the heap if you store a different size than you actually
allocate.

* If a secure_putbyte fails, set nout to zero so you don't run off the
end of the buffer next time around.

* Only write out foure bytes of net_len, no matter how big it is.  The
right answer is to have it be some 32-bit type but I'm not sure if I
should use the krb5 type, the GSSAPI type, or what.  (Remember, this
code has ifdefs for KerberosIV without GSSAPI)

* While we're at it, if secure_write fails while writing out a file in
the client, notice the error. (a break in an inner loop didn't break
out quite far enough)

	With these changes, I am able to get and put multi-megabyte
files even on an Alpha.

git-svn-id: svn://anonsvn.mit.edu/krb5-appl/trunk@2464 dc483132-0cff-0310-8789-dd5450dbe970