Never set krb5_force_static.  Instead, set the defaults for building shared libraries, and allow a Makefile.in to include a new makefile fragment for building private (static, not installed) libraries.
Created another makefile fragment to be included to indicate the
shared library has no dependencies.  (Currently this is the case only
for the libdb2 library, which we don't install, but do build for
testing.)  The way we construct the library dependency search path
arguments doesn't work for an empty list on some platforms.

Updated Makefile.in to use @libpriv_frag@ and @libnodeps_frag@ as
necessary.

git-svn-id: svn://anonsvn.mit.edu/krb5-appl/trunk@3190 dc483132-0cff-0310-8789-dd5450dbe970