move generated dependencies out of Makefile.in
Move automatically-generated dependencies into separate files in the
source tree, and take the data out of Makefile.in.

Keep the "make depend" rules for stripping out the dependencies from
Makefile.in, in case some optional directories were missed, but
everything that builds on my UNIX build has been converted.
(Converting a directory just requires creating an empty "deps" file so
that config.status can build the makefile, and then later running
"make depend" in that directory to get the correct content for it.)

Change configure scripts to incorporate the "deps" file when building
each Makefile.  This change requires the existence of a file "deps" in
each source directory where we build a makefile, even if there are no
sources for which to compute dependencies; a switch to GNU make would
let us conditionalize that, but we can assess that later.

Update dependencies for the generate Makefile itself to list the deps
file.

This will also require some minor tweaking of the Windows build, to
make it incorporate the new deps file.

ticket: new

git-svn-id: svn://anonsvn.mit.edu/krb5-appl/trunk@3239 dc483132-0cff-0310-8789-dd5450dbe970