Portability fixes
HURD does not define MAXPATHNAMELEN nor MAXHOSTNAMELEN.  After a long
discussion with the HURD maintainers they have convinced me this is
correct according to POSIX if no reasonable limit exists.  So, assume
a value if one is not present.

Alpha Linux requires sys/ioctl.h to be included in a couple of files.

git-svn-id: svn://anonsvn.mit.edu/krb5-appl/trunk@3306 dc483132-0cff-0310-8789-dd5450dbe970