Don't limit the length of an input line in FTP
Replace static buffers and fgets() with dynamic buffers and a knockoff
of getline() that we'll call getinput().  Leave the limit on the number
of arguments in place, and make a guess that macro expansion won't
require more than 500 chars beyond the length of the macro definition
and the current command as it was entered.